# malfini-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Eshop",
    "description": "\nAPI Support Contact: b2bsupport@malfini.com\n\nFor more contacts, visit www.malfini.com, section Contacts.\n\nTo authorize in auth or api-auth login functions, use the username and password from the www.malfini.com e-shop. It is recommended to use the Main Account / User Details section in the e-shop to create a new dedicated account for the purpose of accessing the API.\n\nFor the ver parameter, use the current version of the API (4).",
    "version": "4.0"
  },
  "paths": {
...
    "/api/v4/product": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "List of products.",
        "description": "Returns full list of products (nomenclatures) in portfolio, and detailed information about them.",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "description": "ISO 639-1 format.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productCodes",
            "in": "query",
            "description": "List of product code for filtering, separated by commas. If not specified, return all.",
            "schema": {
              "pattern": "[a-zA-Z0-9,]+",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTTP 200 and list of products if OK.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductModel"
                  }
                }
              }
            }
          },
          "500": {
            "description": "HTTP 500 in a case of unexpected error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "HTTP 429 in a case of too many requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "HTTP 401 in a case of not being authorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Cookies": [ ],
            "Bearer": [ ]
          }
        ]
      },
      "head": {
        "tags": [
          "product"
        ],
        "summary": "List of products.",
        "description": "Returns full list of products (nomenclatures) in portfolio, and detailed information about them.",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "description": "ISO 639-1 format.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productCodes",
            "in": "query",
            "description": "List of product code for filtering, separated by commas. If not specified, return all.",
            "schema": {
              "pattern": "[a-zA-Z0-9,]+",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTTP 200 and list of products if OK.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductModel"
                  }
                }
              }
            }
          },
          "500": {
            "description": "HTTP 500 in a case of unexpected error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "HTTP 429 in a case of too many requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "HTTP 401 in a case of not being authorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Cookies": [ ],
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/v4/product/recommended-prices": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Recommended prices.",
        "description": "Returns recommended prices for products, where they are available.",
        "parameters": [
          {
            "name": "productCodes",
            "in": "query",
            "description": "List of product code for filtering, separated by commas. If not specified, return all.",
            "schema": {
              "pattern": "[a-zA-Z0-9,]+",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductRecommendedPriceExternalDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Cookies": [ ],
            "Bearer": [ ]
          }
        ],
        "x-codeSamples": [
          {
            "lang": "PHP",
            "label": "C#",
            "source": "Console.WriteLine(\"Hello World\");"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "Console.WriteLine(\"Hello World\");"
          }
        ]
      },
      "head": {
        "tags": [
          "product"
        ],
        "summary": "Recommended prices.",
        "description": "Returns recommended prices for products, where they are available.",
        "parameters": [
          {
            "name": "productCodes",
            "in": "query",
            "description": "List of product code for filtering, separated by commas. If not specified, return all.",
            "schema": {
              "pattern": "[a-zA-Z0-9,]+",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductRecommendedPriceExternalDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Cookies": [ ],
            "Bearer": [ ]
          }
        ],
        "x-codeSamples": [
          {
            "lang": "PHP",
            "label": "C#",
            "source": "Console.WriteLine(\"Hello World\");"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "Console.WriteLine(\"Hello World\");"
          }
        ]
      }
    },
    "/api/v4/product/prices": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Product prices.",
        "description": "Returns your prices for all products",
        "parameters": [
          {
            "name": "productCodes",
            "in": "query",
            "description": "List of product code for filtering, separated by commas. If not specified, return all.",
            "schema": {
              "pattern": "[a-zA-Z0-9,]+",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductPriceExternalDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Cookies": [ ],
            "Bearer": [ ]
          }
        ],
        "x-codeSamples": [
          {
            "lang": "PHP",
            "label": "C#",
            "source": "Console.WriteLine(\"Hello World\");"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "Console.WriteLine(\"Hello World\");"
          }
        ]
      },
      "head": {
        "tags": [
          "product"
        ],
        "summary": "Product prices.",
        "description": "Returns your prices for all products",
        "parameters": [
          {
            "name": "productCodes",
            "in": "query",
            "description": "List of product code for filtering, separated by commas. If not specified, return all.",
            "schema": {
              "pattern": "[a-zA-Z0-9,]+",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductPriceExternalDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Cookies": [ ],
            "Bearer": [ ]
          }
        ],
        "x-codeSamples": [
          {
            "lang": "PHP",
            "label": "C#",
            "source": "Console.WriteLine(\"Hello World\");"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "Console.WriteLine(\"Hello World\");"
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProductModel": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "The 3-character alphanumeric code of the product.",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "The product's name.",
            "nullable": true
          },
          "categoryName": {
            "type": "string",
            "description": "The name of the category the product belongs to.",
            "nullable": true
          },
          "categoryCode": {
            "type": "string",
            "description": "Unique category ID.",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "description": "The product's gender (or width in case of boots).",
            "nullable": true
          },
          "genderCode": {
            "type": "string",
            "description": "Unique gender ID.",
            "nullable": true
          },
          "trademark": {
            "type": "string",
            "description": "The product's trademark.",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "The product's type.",
            "nullable": true
          },
          "subtitle": {
            "type": "string",
            "description": "The product's subtitle on the e-shop.",
            "nullable": true
          },
          "specification": {
            "type": "string",
            "description": "The product's summary/specification.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "The product's additional description.",
            "nullable": true
          },
          "productCardPdf": {
            "type": "string",
            "description": "Link to the product's product card in the given language.",
            "nullable": true
          },
          "sizeChartPdf": {
            "type": "string",
            "description": "Link to the product's size chart.",
            "nullable": true
          },
          "userInformationPdf": {
            "type": "string",
            "description": "Link to the product's user information file.",
            "nullable": true
          },
          "declarationPdf": {
            "type": "string",
            "description": "Link to the declaration about the product.",
            "nullable": true
          },
          "technicalSpecificationPdf": {
            "type": "string",
            "description": "Link to the product's technical specification.",
            "nullable": true
          },
          "certificationPdf": {
            "type": "string",
            "description": "Link to the product's certification file.",
            "nullable": true
          },
          "additionalInformationPdf": {
            "type": "string",
            "description": "Link to the product's additional information file.",
            "nullable": true
          },
          "alternatives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductAlternativeModel"
            },
            "description": "The alternatives to this product (different genders).",
            "nullable": true
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantModel"
            },
            "description": "The variants the product is available in (for example colors).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Product object"
      },
...
      "ProductRecommendedPriceExternalDto": {
        "type": "object",
        "properties": {
          "productSizeCode": {
            "type": "string",
            "description": "Nomenclature code - 7 A-Z-0-9 characters referring to the commodity, color and size of the product.",
            "nullable": true
          },
          "price": {
            "type": "number",
            "description": "The price value.",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "description": "The currency the price is in.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Cookies": {
        "type": "http",
        "description": "JWT Authorization header using the Cookie scheme",
        "scheme": "Cookies"
      },
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme",
        "scheme": "Bearer"
      }
    }
  }
}